debug

+++ src/radix_r_bruck.cpp
@@ -109,7 +109,7 @@ void uniform_radix_r_bruck(std::vector<int>& act_sd_pstep, int r, char *sendbuf,
 //    		e = MPI_Wtime();
 //    		pre_time += e - s;
 
-    		int distance = z * pow(r, x);
+    		int distance = (int)z * pow(r, x);
     		if (rank == 0)
     			std::cout << x << " " << z << " " << distance << " " << pow(r, x) << " " << pow(static_cast<float>(r), static_cast<float>(x)) << std::endl;
